完成本地上传文件功能

+++ src/main/java/com/dimple/project/tool/service/impl/LocalStorageServiceImpl.java
@@ -4,11 +4,11 @@ import com.dimple.common.exception.CustomException;
 import com.dimple.common.utils.SecurityUtils;
 import com.dimple.common.utils.StringUtils;
 import com.dimple.common.utils.file.FileUtils;
+import com.dimple.framework.config.DimpleBlogConfig;
 import com.dimple.project.tool.domain.LocalStorage;
 import com.dimple.project.tool.mapper.LocalStorageMapper;
 import com.dimple.project.tool.service.LocalStorageService;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
@@ -28,9 +28,6 @@ public class LocalStorageServiceImpl implements LocalStorageService {
     @Autowired
     LocalStorageMapper localStorageMapper;
 
-    @Value("${file.path}")
-    private String path;
-
     @Override
     public List<LocalStorage> selectLocalStorageList(LocalStorage localStorage) {
         return localStorageMapper.selectLocalStorageList(localStorage);
@@ -42,15 +39,14 @@ public class LocalStorageServiceImpl implements LocalStorageService {
         //获取后缀
         String suffix = FileUtils.getExtensionName(multipartFile.getOriginalFilename());
         String type = FileUtils.getFileType(suffix);
-        File file = FileUtils.upload(multipartFile, path + type + File.separator);
+        File file = FileUtils.upload(multipartFile, DimpleBlogConfig.getProfile() + type + File.separator);
         if (Objects.isNull(file)) {
             throw new CustomException("上传失败");
         }
         //防止异常出错
         try {
             name = StringUtils.isBlank(name) ? FileUtils.getFileNameNoExtension(multipartFile.getOriginalFilename()) : name;
-            LocalStorage localStorage = new LocalStorage(file.getName(), name, suffix, file.getPath(), type, FileUtils.getSizeString(multipartFile.getSize())
-            );
+            LocalStorage localStorage = new LocalStorage(file.getName(), name, suffix, file.getPath(), type, FileUtils.getSizeString(multipartFile.getSize()));
             localStorage.setCreateBy(SecurityUtils.getUsername());
             return localStorageMapper.insertLocalStorage(localStorage);
         } catch (Exception e) {
@@ -68,6 +64,13 @@ public class LocalStorageServiceImpl implements LocalStorageService {
     @Override
     public int deleteLocalStorage(Long id) {
         String username = SecurityUtils.getUsername();
+        LocalStorage localStorage = localStorageMapper.selectLocalStorageById();
+        if (Objects.isNull(localStorage)) {
+            throw new CustomException("文件不存在");
+        }
+        //删除文件
+        String path = localStorage.getPath();
+        FileUtils.del(path);
         return localStorageMapper.deleteLocalStorageById(id, username);
     }
 }
